Add and update user (buaya)

+++ src/main/java/com/buaya/security/model/Handicap.java
@@ -17,28 +17,28 @@ public class Handicap implements Serializable{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private long id;
 	
 	@Column(name="user_id")
-	private int userId;
+	private long userId;
 	
 	@Column(name="handicap_value")
 	private long handicapValue;
 	
 	public Handicap() {}
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
-	public int getUserId() {
+	public long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
