+++ src/main/java/com/buaya/security/repository/HandicapRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.buaya.security.model.Handicap;
 
-public interface HandicapRepository extends JpaRepository<Handicap, Integer>{
+public interface HandicapRepository extends JpaRepository<Handicap, Long>{
 	
-	public Optional<Handicap> findByUserId(int userId);
+	public Optional<Handicap> findByUserId(long userId);
 	
 }
